musl: prevent symlink for dynamic linker being installed in /lib
it seems a recent musl update caused the dynlinker to be installed to /lib
when running make install (without DESTDIR), regardless of --prefix.
this would either cause build to break as non-root, or worse, as root
overwrite existing symlinks for the dynlinker, pointing into the cross-
root, which would cause breakage when the cross-root directory is removed.

we fix this by building musl with --prefix= and then
make install DESTDIR=$PREFIX.