add preliminary support for GCC 4.4.7
patches are forward-and-backwards-ports of the MCM patches for GCC 4.2.1
and 5.3.0 respectively and identical in functionality.
this makes it the first compiler patch series in musl-cross that got
even the sub-arch dynlinker identification bits (such as -eb/-le/-sf...)
right. only the musl-compatibility bits have been ported so far, plus
one build fix. whether other patches from MCM should be adopted remains
to be evaluated.

the following archs built successfully: x86_64, i486, powerpc, mips, armv7.
both with and without C++ support.
sh4 fails even though the forward-ported dynlinker patches apply cleanly;
but sh4 is known to be broken on all GCCs pre-4.8.5.
they were also tested to produce working binaries with the right dynlinker
names.

GCC 4.4.7 is of interest in so far as it's the last pre-LTO version, with
the register allocator bugs of 4.2.x series removed, and apparently a pretty
well-balanced code generator[0].

stripped C-only toolchains with manpages removed take about 35 MB of disk
space, which is only ~5MB bigger than 4.2.x series.

with this patch done, it should be relatively straight-forward to get 4.3.x
or 4.5.x working too.

[0]: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=51017