bump binutils default version to 2.27
binutils patches taken from musl-cross-make master as of today.

binutils 2.27, which was the default in mcm, has proven itself over time
and seems to work well with all supported archs, just like 2.24.90, but
it supports some features newer binutils default to, which 2.24.90 can't
deal with (this breaks when you link e.g. to some libraries created with
newer binutils).