implement fetching of config.sub/.guess
this fixes issues when building on a platform that isn't already
in config.guess with older GCCs, such as e.g. on aarch64.
additionally, more triples like armv[0-8] are now properly detected.
as a bonus, future musl patches don't need the config.sub patches
either.
(we copy the files in place after patching, so the old patches keep
working.)